// schemas/reservation.schema.ts
import { z } from "zod";

export const ReservationSchema = z.object({
    id: z.number(),
    initial_listing_id: z.number().nullable(),
    currency: z.string(),
    price_per_night: z.number(),
    base_price: z.number(),
    security_price: z.number().nullable(),
    extras_price: z.number().nullable(),
    cleaning_fee: z.number(),
    channel_commission: z.number().nullable(),
    service_charge: z.number().nullable(),
    subtotal: z.number(),
    payout_price: z.number(),
    transaction_fee: z.number().nullable(),
    sum_refunds: z.number().nullable(),
    revenue: z.number(),
    owner_revenue: z.number(),
    net_revenue: z.number(),
    channel_reservation_id: z.number(),
    source: z.string(),
    status_code: z.number(),
    status: z.string(),
    status_description: z.string(),
    confirmation_code: z.string(),
    nights: z.number(),
    checkIn: z.string(),
    checkOut: z.string(),
    guests: z.number(),
    adults: z.number(),
    children: z.number().nullable(),
    infants: z.number().nullable(),
    confirmed_at: z.string(),
    advance_days: z.number().nullable(),
    cancel_policy: z.any().nullable(),
    planned_arrival: z.any().nullable(),
    planned_departure: z.any().nullable(),
    beds_to_be_prepared: z.any().nullable(),
    notes: z.string().nullable(),
    cleaning_notes: z.string().nullable(),
    updated_at: z.string(),
    cancelled_at: z.string().nullable(),
    is_manual: z.number(),
    added_by: z.any().nullable(),
    extra_info: z.any().nullable(),
    cancel_penalty: z.any().nullable(),
    guest_details: z.any().nullable(),
    price_details: z.any().nullable(),
    addons_price: z.number().nullable(),
    addons_details: z.any().nullable(),
    city_tax: z.number().nullable(),
    extras_price_ex: z.number().nullable(),
    extras_price_inc: z.number().nullable(),
    cancellation_fee: z.number().nullable(),
    checked_in_mark: z.any().nullable(),
    checked_out_mark: z.any().nullable(),
    payout_by_channel: z.number(),
    payout_by_channel_changed: z.number(),
    pre_check_in_completed: z.number(),
    pre_check_out_completed: z.number(),
    document_uploaded: z.number(),
    passport_uploaded: z.number(),
    rate_name: z.string().nullable(),
    created_at: z.string().nullable(),
    message_id: z.number(),
    discount_code: z.string(),
    discount_id: z.number(),
    paid_part: z.string(),
    paid_sum: z.number().nullable(),
    parent_listing_id: z.number().nullable(),
    listing_id: z.number(),
    integration_id: z.number(),
    guest_id: z.number(),
    inbox_id: z.any().nullable(),
    tax_amount: z.number()
});

export type ReservationType = z.infer<typeof ReservationSchema>;